+++ lib/platform/platform_features_web.dart
@@ -1,5 +1,7 @@
 bool get isDesktopNative => false;
 
+bool get isLinuxDesktop => false;
+
 bool get supportsNativeGit => false;
 
 bool get supportsDualScreenPresenter => false;